+++ packages/assets-ergot/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-assets-ergot
 
+## 0.51.13
+
+### Patch Changes
+
+- @happyvertical/smrt-assets@0.51.13
+
+## 0.51.12
+
+### Patch Changes
+
+- @happyvertical/smrt-assets@0.51.12
+
 ## 0.51.11
 
 ### Patch Changes