+++ packages/personas/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @happyvertical/smrt-personas
 
+## 0.51.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.13
+  - @happyvertical/smrt-agents@0.51.13
+  - @happyvertical/smrt-messages@0.51.13
+  - @happyvertical/smrt-prompts@0.51.13
+  - @happyvertical/smrt-tenancy@0.51.13
+  - @happyvertical/smrt-users@0.51.13
+  - @happyvertical/smrt-ui@0.51.13
+
+## 0.51.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.12
+  - @happyvertical/smrt-agents@0.51.12
+  - @happyvertical/smrt-messages@0.51.12
+  - @happyvertical/smrt-prompts@0.51.12
+  - @happyvertical/smrt-tenancy@0.51.12
+  - @happyvertical/smrt-users@0.51.12
+  - @happyvertical/smrt-ui@0.51.12
+
 ## 0.51.11
 
 ### Patch Changes