+++ packages/smrt-workbench/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @happyvertical/smrt-workbench
 
+## 0.51.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.13
+  - @happyvertical/smrt-svelte@0.51.13
+  - @happyvertical/smrt-playground@0.51.13
+  - @happyvertical/smrt-ui@0.51.13
+
+## 0.51.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.12
+  - @happyvertical/smrt-svelte@0.51.12
+  - @happyvertical/smrt-playground@0.51.12
+  - @happyvertical/smrt-ui@0.51.12
+
 ## 0.51.11
 
 ### Patch Changes