+++ packages/vitest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @happyvertical/smrt-vitest
 
+## 0.51.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.13
+
+## 0.51.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.51.12
+
 ## 0.51.11
 
 ### Patch Changes